Rename XML division parser module to parse_divs
